Progress with min dwell times

+++ TrainUnit.h
@@ -130,8 +130,8 @@ public:
 ///< indicates a train that is defined by the timetable as under signaller control
     bool Warning;
 ///< if set triggers an alert in the warning and perf log panels when the action is reached
-    int MinDwellTime; //added after v2.22.0
-///< minimum waiting time at a location, default 30s
+    double MinDwellTime; //added after v2.22.0
+///< minimum waiting time at a location, default 30s, float because float needed for TDateTime
     int NumberOfRepeats;
 ///< the number of repeating services
     int RearStartOrRepeatMins, FrontStartOrRepeatDigits;
@@ -175,7 +175,7 @@ public:
         Warning = false;
         Reminder = 0;
         SignallerControl = false;
-        MinDwellTime = 30; //default value in seconds
+        MinDwellTime = 30.0; //default value in seconds
     }
 };
